revert: move add account to addAccounts machine
