+++ packages/app/src/systems/Account/machines/addAccountMachine.test.ts
@@ -13,8 +13,6 @@ const machine = addAccountMachine.withContext({}).withConfig({
   },
 });
 
-const name = 'Account Go';
-
 describe('addAccountMachine', () => {
   let service: AddAccountMachineService;
 
@@ -30,32 +28,13 @@ describe('addAccountMachine', () => {
   describe('add', () => {
     it('should be able to add an account', async () => {
       await expectStateMatch(service, 'idle');
-      service.send('ADD_ACCOUNT', {
-        input: name,
-      });
+      service.send('ADD_ACCOUNT');
 
       await expectStateMatch(service, 'addingAccount');
       await expectStateMatch(service, 'idle');
 
       const accounts = await AccountService.getAccounts();
-      expect(accounts?.[1].name).toBe(name);
-    });
-
-    it('should not be able to add accounts with same name', async () => {
-      await expectStateMatch(service, 'idle');
-      service.send('ADD_ACCOUNT', {
-        input: name,
-      });
-      await expectStateMatch(service, 'addingAccount');
-      await expectStateMatch(service, 'idle');
-      service.send('ADD_ACCOUNT', {
-        input: name,
-      });
-
-      // make sure test fails but jest don't stop
-      jest.spyOn(console, 'error').mockImplementation();
-
-      await expectStateMatch(service, 'failed');
+      expect(accounts?.[1].name).toBe(`Account 2`);
     });
   });
 });
